Initialize DefaultPackageItem.containingPackageField
Previously, `DefaultPackageItem.containingPackageField` was a
`lateinit var` that was only initialized on demand. This change adds a
`containingPackage` constructor parameter that is used to initialize
it. The parameter is optional at the moment as not all creators of
`DefaultPackageItem`s pass in a `containingPackage`. If it is not
specified then it defaults to `null` in which case it falls back to the
previous behavior.

Follow-up changes will update the creators and will then make the
`containingPackage` mandatory and remove any unnecessary code.

Bug: 352480646
Test: ./gradlew
Change-Id: Ibb278d7b1630cd305f2dc46fc03b37ab336930fa