Replace mutableModifiers() with mutateModifiers {...}
Previously, the `mutableModifiers()` would return a refence to the
`Item`'s modifiers for direct modification. The life time of the
returned object was unbounded so there was no way to change the
underlying representation to be immutable with copy-on-mutate
semantics.

This change replaces `mutableModifiers()` with `mutateModifiers {...}`
which takes a lambda that is called on a `MutableModifierList` object.
While the current implementation still provides access to the
`Item`'s modifiers it does allow future changes to implement
copy-on-mutate semantics and limit mutations to within the lambda of
the `mutateModifiers {...}` method.

Bug: 356830333
Test: ./gradlew
Change-Id: Ia8bde8bca17c1f70551875cdde55cef0d2a2f60d