Avoid mutating modifiers in createPackageItem(...)
Although packages do not have modifiers (only annotations) internally
they do have modifiers and in order to ensure consistent behavior they
are expected to be public. Previously, that was enforced by
`createPackageItem(...)` mutating the modifiers that it was provided to
explicitly set it. This change avoids that mutation by ensuring that
the callers explicitly set it to be public (which they do).

The default value for modifiers was wrong as it was `PACKAGE_PRIVATE`
but rather than make that `PUBLIC` this change just removes it (and all
the other defaults) as the sole caller always provides all the
parameters.

Bug: 356830333
Test: ./gradlew
Change-Id: Ia0b9a8f874f8bc942a93bf903c06b50c5ee7700a