Add DefaultModifierList(VisibilityLevel, List<AnnotationItem)
One of the issues with `DefaultModifierList` is that it exposes quite a
bit of its internal implementation details to users. e.g. when creating
a new instance it is necessary to pass in one of the internal flag
values to set the visibility level.

This change adds a new constructor that takes a `VisibilityLevel`
instead of a `flags: Int`. It also changes all existing calls that do
not provide a full set of flags to use this. The flags constructor is
only used by Psi and Turbine  who pre-calculate the `flags`. They
always provide flags so the `flags` constructor parameter did not need
a default value so it was removed.

Bug: 356830333
Test: ./gradlew
Change-Id: Icabc4f58e9c6e7bbf2179218325512ae58fc752c