Separate ModifierList and MutableModifierList
Previously, `MutableModifierList` extended `ModifierList` which was a
problem as that would allow a `MutableModifierList` to be passed to
some code that expected an immutable `ModifierList` and then change it
afterwards. If the seemingly immutable `ModifierList` was cached for
re-use then that would lead to some very hard to debug issues.

Mutating the modifiers after storing in `Item.modifiers` is actually
something that does happen so this is not a theoretical issue.
Fortunately, it does not cause any harm because up to now
`Item.modifiers` has been considered to be mutable but that is
changing so these issues will be addressed in follow-up changes.

This changes creates a new structure like this:
* `BaseModifierList` - contains common code for accessing contents. It
  is basically what `ModifierList` was.
* `ModiferList` extends `BaseModifierList` but is otherwise empty. This
  is the immutable set of modifiers and annotations.
* `MutableModifierList` extends `BaseModifierList` instead of
  `ModifierList`.
* `DefaultModifierList` implements both `MutableModifierList` and
  `ModifierList` temporarily to ease migration.

Follow-up changes will start by changing uses of `DefaultModifierList`
to the appropriate interface and once that has been done then the
`DefaultModifierList` will be split into similarly organized classes.

Bug: 356830333
Test: ./gradlew
Change-Id: I487face9ae1e113f830401fd1f7e984b8c2ab6b8