Change AbstractItem.modifiers type to ModifierList
Previously, while the `Item.modifiers` property is of type
`ModifierList` the `AbstractItem.modifiers` property was of type
`DefaultModifierList`. The discrepancy made some refactoring difficult
if code that had access to `modifiers: DefaultModifierList` only had
access to `modifiers: ModifierList`. Also, on a more general note
`DefaultModifierList` is an implementation detail of `ModifierList`
and `MutableModifierList` and so should not be being used in all the
places it currently is.

Over the next few changes the uses of `DefaultModifierList` will be
replaced with one of the interfaces it implements as follows:
1. `AbstractItem.modifiers` will be changed to `ModifierList` to match
   `Item.modifiers`.
2. Code that currently creates `DefaultModifierList` instances directly
   will create `MutableModifierList` or `MutableList` instances as
   needed.
3. Code that can accept either will use `BaseModifierList`.

This change switches the `AbstractItem.modifiers` to `ModifierList` as
that must be immutable. However, the `modifiers` constructor parameters
use `BaseModifierList` so they can accept both `MutableModifierList`
and `ModifierList` making it easier for callers.

Mapping from the `BaseModifierList` parameter to the `ModifierList`
property is done by calling the `BaseModifierList.toImmutable()`
method added as part of this change. `DefaultModifierList` implements
that simply by returning itself but a future change will create a copy
if necessary.

Changing `AbstractItem.modifiers` to `ModifierList` required that the
`mutateModifiers {...}` method was changed too as it no longer had
access to a `DefaultModifierList`. Rather than cast to a
`MutableModifierList` this change uses the new `toImmutable()` and a
complementary `toMutable()` method to implement the copy-on-mutation
semantics, even though for the moment there is no copying done.

Bug: 356830333
Test: ./gradlew
Change-Id: Ia125729c727a80482e25a402e0c3c423ec36c499