Optimize ModifierList.snapshot()
Previously, it was defined on `BaseModifierList` but it is only called
on `ModifierList`. So, this moves it down to there and moves the
implementation down to `DefaultModifierList`. As that class is
immutable as long as there are no annotations then `snapshot()` can
just return `this` as it is safe to share `DefaultModifierList`s
without any annotations across `Item`s even if they come from different
`Codebase`s.

Bug: 356830333
Test: ./gradlew
Change-Id: If01bd903f5890f18bd69d5e1d1e3417d8a490767