Flip Codebase/Assembler creation order
The `DefaultCodebase` and `CodebaseAssembler` objects are tightly
coupled as they refer to each other. In order to initialize the
references without having to use mutable properties they need to be
created together. Previously, `DefaultCodebase` was passed an
`assemblerFactory` that it could use in its initialization code to
create a `CodebaseAssembler`, passing in itself as the `codebase`
parameter.

That structure was used for a couple of reasons:
1. The model specific `Codebase` implementations like `TextCodebase`
   and `TurbineBasedCodebase` were the primary focus during
   construction and    `CodebaseAssembler` was peripheral.
2. It required less casting as users of `CodebaseAssembler` did not
   need to access the model specific implementation but they did need
   to access the model specific `Codebase` implementation.

Since then a lot of work has been done to move functionality from the
model specific `Codebase` implementations into the `CodebaseAssembler`
imeplementations and the `TextCodebase` and `TurbineBasedCodebase`
implementations have been removed. So, the previous reasons no longer
hold.

This change flips the order by passing the `CodebaseAssembler`
implementations a `codebaseFactory` parameter that they can use to
construct the `DefaultFactory`, passing in themselves as the
`assembler` parameter.

Bug: 352480646
Test: ./gradlew
Change-Id: I7fb6c976e659f0d7704efc080577d3de3cab3bff