SignatureWriter: Replace filterEmit/Reference with ApiType
Previously, all callers effectively created `filterEmit` and
`filterReference` from an `ApiType` passing in an
`ApiPredicate.Config`. Those that did not do that explicitly provided
`{ true }` for both which is effectively using `ApiType.ALL`.

This change dedups that code by replacing the `filterEmit` and
`filterReference` with an `ApiType` and moving the creation of those
predicates inside the `createFilteringVisitor(...)` method. It does
not need to pass in an `ApiPredicate.Config` as `apiVisitorConfig`'s
`apiPredicateConfig` is identical to that previously provided.

Bug: 357833584
Test: ./gradlew
Change-Id: I207c6cb105d7a180876c874025c06a32adbf64de