Generate signatures using snapshot including reverted items
Previously, `SignatureWriter` would use modifiers from previously
released API for any API changes that have been reverted. This removes
that special handling and instead creates a snapshot that combined
un-reverted items from the sources and reverted items from the
previously released API.

This does not handle documentation correctly, it will drop
documentation from reverted items. That is not a problem for signature
files but will be an issue for stub files so it will be fixed in a
follow-up change.

Bug: 357833584
Test: ./gradlew
Change-Id: I42232b97882a628d3dc39efa497ad8a1c51d2413