Fix recursive call to TextCodebaseAssembler.getOrCreateClass(...)
Previously, the recursive call would always specify
`isOuterClassOfClassInThisCodebase = true`. However, while it did not
cause any issues it was incorrect, e.g. assume that a request was made
for `java.lang.Character.Subset` and not `ClassResolver` was provided.
It would compute a `fullName` of `Character.Subset` and then because
that contained a `.` it would look for `java.lang.Character` but it
would assume that was an outer class of a class in the codebase but
it would not be true.

This change fixes it to just pass `isOuterClassOfClassInThisCodebase`
through, so that only outer classes of classes defined in a signature
file would have `isOuterClassOfClassInThisCodebase = true`. This is
needed because a follow-up change will use the value of that
parameter to control the `emit` property of the containing package and
so it is important that it is correct.

Bug: 357833584
Test: ./gradlew
Change-Id: I84419a8f0a2d6ca10c713856b3b2bd32f48afd9f