Pass SourceFile through to the snapshot
There seems little advantage to creating a special snapshot class of
`SourceFile` at the moment. It would require mapping the model specific
types underlying the imports into `Item` specific types and then
reimplementing the import filtering on top of that. While that would be
a good idea in the longer term it seems like quite a bit of work for
little benefit right now.

So, this change simply passes the `SourceFile`, if any, through to the
snapshot.

Bug: 353737744
Test: ./gradlew
Change-Id: I243ed29a8a6a6ae362b87f268ea2ed9d8f91928d