Support testing against multiple SourceModelProviders
Previously, the `SourceModelSuiteRunner` only supported a single
`SourceModelProvider`. That meant that it was not possible for the
`metalava-model-snapshot-testing` project to test against Turbine and
Psi. Rather than try and make `SourceModelSuiteRunner` support multiple
`SourceModelProvider`s this change adds a new service interface
`ModelSuiteRunnerProvider` that allows multiple `ModelSuiteRunner`
instances to be returned.

The `SourceModelSuiteRunnerProvider` can then just return one
`SourceModelSuiteRunner` for each `SourceModelProvider` it finds.

A follow-up change will run `metalava-model-snapshot-testing` against
`metalava-model-psi` as well as `metalava-model-turbine` and
`metalava-model-text` using this.

Bug: 353737744
Test: ./gradlew
Change-Id: I56cd9232dee9537fe6e90850ced43023f761c411